fix(xrpl): gosec linting

+++ xrpl/transaction/payment_channel_fund.go
@@ -68,7 +68,7 @@ func (p *PaymentChannelFund) Validate() (bool, error) {
 
 	// check the expiration time is in the future. /!\ Incomplete as the channel SettleDelay is not taken into account but it's already a good check.
 	currentRippleTime := rippletime.UnixTimeToRippleTime(time.Now().Unix())
-	if (p.Expiration != 0) && (p.Expiration < uint32(currentRippleTime)) {
+	if (p.Expiration != 0) && (int64(p.Expiration) < currentRippleTime) {
 		return false, ErrInvalidExpiration
 	}
